refactor: 채팅 input 컴포넌트 분리 및 쿼리캐시 설정 조정 #134
